Make `PieceIndex` a new type with helpful methods to reduce usage of `PIECES_IN_SEGMENT`

+++ crates/sp-lightclient/src/tests.rs
@@ -149,7 +149,7 @@ impl PieceGetter for TestPieceGetter {
             .archived_segment
             .pieces
             .iter()
-            .nth(piece_index as usize)
+            .nth(u64::from(piece_index) as usize)
             .map(Piece::from))
     }
 }
